requirements.txt
av==12.0.0
datasets==3.2.0
easydict==1.13
fastprogress==1.0.3
gdown==5.2.0
natsort==8.4.0
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
protobuf==6.30.2
tabulate==0.9.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.51.0
